refactor: consolidate interpolateVars functionality and remove duplicate interpolateVars files
